Share common downloader
